requirements.txt
ace_tools==0.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
deprecation==2.1.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.13.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.2.0
idna==3.10
ieeg==1.6
imbalanced-learn==0.13.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lightning-utilities==0.11.9
line_profiler==4.1.3
MarkupSafe==2.1.5
matplotlib==3.9.3
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==9.1.0.70
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.21.5
# nvidia-nvjitlink-cu12==12.1.105
# nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pennprov==2.2.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyEDFlib==0.1.40
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
# pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
scikit-learn==1.6.0
scikit-posthocs==0.11.4
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sklearn-compat==0.1.3
stack-data==0.6.3
statannotations==0.7.1
statsmodels==0.14.4
sympy==1.13.1
threadpoolctl==3.5.0
# torch==2.5.1+cu121
# torchaudio==2.5.1+cu121
# torchmetrics==1.6.0
# torchvision==0.20.1+cu121
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
xgboost==2.1.4
yarl==1.18.0
